test: reproduce terminal automatic top inset
